+++ jest.config.js
@@ -5,6 +5,7 @@ module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
   watchPathIgnorePatterns: ['<rootDir>/examples/'],
+  testPathIgnorePatterns: ['<rootDir>/__tests__/mocks/'],
   moduleNameMapper: pathsToModuleNameMapper(compilerOptions.paths, {
     prefix: '<rootDir>/'
   })